requirements.txt
streamlit
openai
python-dotenv
supabase